+++ extensions/mailersend/src/domains.tsx
@@ -51,23 +51,41 @@ function Activities({ domain }: { domain: Domain }) {
     [ActivityEventType.SPAM_COMPLAINTS]: undefined,
   };
 
-  
   const {
     isLoading,
     data: activities,
     pagination,
   } = useMailerSendPaginated<Activity>(`activity/${domain.id}?date_from=${from}&date_to=${to}`);
-  
-  const filteredActivities = !filter ? activities : activities.filter(a => a.type===filter);
-  
+
+  const filteredActivities = !filter ? activities : activities.filter((a) => a.type === filter);
+
   return (
-    <List isLoading={isLoading} pagination={pagination} searchBarAccessory={<List.Dropdown tooltip="Event status" onChange={setFilter}>
-      <List.Dropdown.Item icon={Icon.Dot} title="All" value="" />
-      <List.Dropdown.Section>
-      {Object.entries(ActivityEventType).map(([key, val]) => <List.Dropdown.Item key={key} icon={{ source: Icon.Dot, tintColor: TYPE_TO_COLOR[val] }} title={key} value={val} />)}
-      </List.Dropdown.Section>
-    </List.Dropdown>}>
-      {!isLoading && !filteredActivities.length && <List.EmptyView icon={Icon.MagnifyingGlass} title="No activity found" description="Please try again with other keywords, filters or set a different period." />}
+    <List
+      isLoading={isLoading}
+      pagination={pagination}
+      searchBarAccessory={
+        <List.Dropdown tooltip="Event status" onChange={setFilter}>
+          <List.Dropdown.Item icon={Icon.Dot} title="All" value="" />
+          <List.Dropdown.Section>
+            {Object.entries(ActivityEventType).map(([key, val]) => (
+              <List.Dropdown.Item
+                key={key}
+                icon={{ source: Icon.Dot, tintColor: TYPE_TO_COLOR[val] }}
+                title={key}
+                value={val}
+              />
+            ))}
+          </List.Dropdown.Section>
+        </List.Dropdown>
+      }
+    >
+      {!isLoading && !filteredActivities.length && (
+        <List.EmptyView
+          icon={Icon.MagnifyingGlass}
+          title="No activity found"
+          description="Please try again with other keywords, filters or set a different period."
+        />
+      )}
       <List.Section
         title={domain.name}
         subtitle={`${dayjs.unix(from).format("lll")} - ${dayjs.unix(to).format("lll")}`}