+++ extensions/mailersend/src/open-in-mailersend.tsx
@@ -1,6 +1,11 @@
 import { Action } from "@raycast/api";
 
-export default function OpenInMailerSend({route}: {route: string}) {
-    // eslint-disable-next-line @raycast/prefer-title-case
-    return <Action.OpenInBrowser icon="symbol-512.png" title="Open in MailerSend" url={`https://app.mailersend.com/${route}`} />
+export default function OpenInMailerSend({ route }: { route: string }) {
+  return (
+    <Action.OpenInBrowser
+      icon="symbol-512.png"
+      title="Open in Mailersend"
+      url={`https://app.mailersend.com/${route}`}
+    />
+  );
 }